DAOS-3800 dtx: optimize read if hit non-committed sync mode DTX
If the modifications corsses multiple redundancy groups, then it
is possible that the sub modifications on the DTX leader are not
the same as the ones on non-leaders. Under such case, if someone
wants to read the data on some non-leader but hits non-committed
DTX, then asking the client to retry with leader maybe not help.

Instead, we can ask make the client to retry the read again (and
again) sometime later. We do sync commit the DTX with 'DTE_BLOCK'
flag. So it will not cause the client to retry read for too many
times unless such DTX hit some trouble (such as client or server
failure) that may cause current readers to be blocked until such
DTX has been handled by the new leader via DTX recovery.

Currently, we will handle reading EC object similarly since the
DTX leader for EC modification (a parity node) may does not has
the data for client read.

Signed-off-by: Fan Yong <fan.yong@intel.com>